Fixup py27
